requirements.txt
pandas
numpy
requests
schedule
fastapi
uvicorn
streamlit
yfinance
plotly
python-dotenv
polygon-api-client